packages/cv/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "canopy-cv"
version = "0.1.0"
description = "Computer vision pipeline for Canopy Sight — YOLOv9 detection, SAM2 segmentation, and model export"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "torch>=2.3.0",
    "ultralytics>=8.2.0",
    "timm>=1.0.0",
    "albumentations>=1.4.0",
    "pytorch-lightning>=2.2.0",
    "wandb>=0.16.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.17.0",
    "numpy",
    "pillow",
    "opencv-python-headless",
    "pyyaml",
]

[project.optional-dependencies]
sam2 = ["segment-anything-2"]
tensorrt = ["tensorrt"]
dev = ["pytest", "ruff", "mypy"]

[project.scripts]
canopy-train = "canopy_cv.scripts.train:main"
canopy-infer = "canopy_cv.scripts.infer:main"
canopy-export = "canopy_cv.scripts.export_model:main"

[tool.hatch.build.targets.wheel]
packages = ["src/canopy_cv"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
